initial function calling
